requirements.txt
google-api-python-client
requests
pandas
numpy
scikit-learn
Pillow
tqdm
joblib
Flask
python-dotenv
